Update wrapt to 1.16.0
