+++ newrelic/packages/wrapt/__init__.py
@@ -1,11 +1,14 @@
-__version_info__ = ('1', '14', '1')
+__version_info__ = ('1', '16', '0')
 __version__ = '.'.join(__version_info__)
 
-from .wrappers import (ObjectProxy, CallableObjectProxy, FunctionWrapper,
-        BoundFunctionWrapper, WeakFunctionProxy, PartialCallableObjectProxy,
-        resolve_path, apply_patch, wrap_object, wrap_object_attribute,
+from .__wrapt__ import (ObjectProxy, CallableObjectProxy, FunctionWrapper,
+        BoundFunctionWrapper, PartialCallableObjectProxy)
+
+from .patches import (resolve_path, apply_patch, wrap_object, wrap_object_attribute,
         function_wrapper, wrap_function_wrapper, patch_function_wrapper,
         transient_function_wrapper)
+
+from .weakrefs import WeakFunctionProxy
 
 from .decorators import (adapter_factory, AdapterFactory, decorator,
         synchronized)
